backend-production/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
requests==2.32.3
google-generativeai==0.8.3
python-dotenv==1.0.0
pydantic==2.10.5
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
responses==0.25.3
pydantic-settings